requirements.txt
-f /usr/share/pip-wheels
Django==2.2.2
django-crispy-forms==1.7.2
django-js-asset==1.2.2
mysqlclient==1.4.2.post1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
